Refactoring Signed-off-by: Kuldeep Kumar <kuldeep.kumar3@dell.com>
