+++ segmentstore/storage/src/main/java/io/pravega/segmentstore/storage/chunklayer/EpochInfo.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.pravega.segmentstore.server.containers;
+package io.pravega.segmentstore.storage.chunklayer;
 
 import io.pravega.common.ObjectBuilder;
 import io.pravega.common.io.serialization.RevisionDataInput;
